improved speed slightly

+++ heuristics/heuristic_chisquare.jl
@@ -36,15 +36,15 @@ function CHISQUAREHEUR(trans::SparseMatrixCSC{Bool, Int64},
         ## Calculate the weight of each warehouse. It shows us the density of 
         ## the independent coappearances in each warehouse if we were to allocate
         ## all SKUs simply according to the highest independent coappearances.
-        weight = WHWEIGHT(capacity::Array{Int64,1},sum_nor::Array{Float32,1})
+        weight = WHWEIGHT(capacity::Array{Int64,1},sum_nor::Array{Float64,1})
 
         ## Copy the capacity to have an array that keeps the left-over capacity
         cap_left = copy(capacity)
 
         ## Create an array that saves the current dependencies for all unallocated
         ## SKUs to all warehouses and a vector that holds all unallocated skus
-        state_dep = Matrix{Float32}(undef,size(X,1),length(cap_left)) .= 0
-        state_nor = Matrix{Float32}(undef,size(X,1),length(cap_left)) .= 0
+        state_dep = Matrix{Float64}(undef,size(X,1),length(cap_left)) .= 0
+        state_nor = Matrix{Float64}(undef,size(X,1),length(cap_left)) .= 0
         allocated = Vector{Bool}(undef,size(X,1)) .= false
 
         ## In the heuristic we compare the split-delivery minimising potential 
@@ -72,20 +72,20 @@ function CHISQUAREHEUR(trans::SparseMatrixCSC{Bool, Int64},
         ## allocate all significant SKUs from a SKU-cluster into one warehouse. 
         ## Otherwise we allocate it to warehouse k to maximise the independent 
         ## coappearances in the allocation.
-            i,k  = SELECTIK(sum_dep::Vector{Float32},
-                            sum_nor::Vector{Float32},
+        i,k  = SELECTIK(sum_dep::Vector{Float64},
+                            sum_nor::Vector{Float64},
                             weight,
                             cap_left::Array{Int64,1},
                             X::Array{Bool,2},
-                            dep::Matrix{Float32},
+                            dep::Matrix{Float64},
                             allocated::Vector{Bool})
-               ALLOCATEONE!(X::Array{Bool,2},
-                            dep::Matrix{Float32},
-                            nor::Matrix{Float32},
-                            sum_dep::Vector{Float32},
-                            sum_nor::Vector{Float32},
-                            state_dep::Matrix{Float32},
-                            state_nor::Matrix{Float32},
+        ALLOCATEONE!(X::Array{Bool,2},
+                            dep::Matrix{Float64},
+                            nor::Matrix{Float64},
+                            sum_dep::Vector{Float64},
+                            sum_nor::Vector{Float64},
+                            state_dep::Matrix{Float64},
+                            state_nor::Matrix{Float64},
                             cap_left::Array{Int64,1},
                             allocated::Vector{Bool},
                             i::Int64,
@@ -96,17 +96,17 @@ function CHISQUAREHEUR(trans::SparseMatrixCSC{Bool, Int64},
         ## to. If so, check whether the dependencies are expected to dominate the 
         ## independent coapperances. If yes, allocate the corresponding SKUs to 
         ## the warehouse k.
-            ADDDEPENDENT!(X::Array{Bool,2},
+        ADDDEPENDENT!(X::Array{Bool,2},
                           cap_left::Array{Int64,1},
                           k::Int64,
-                          dep::Matrix{Float32},
-                          nor::Matrix{Float32},
-                          sum_dep::Array{Float32,1},
-                          sum_nor::Array{Float32,1},
-                          state_dep::Matrix{Float32},
-                          state_nor::Matrix{Float32},
+                          dep::Matrix{Float64},
+                          nor::Matrix{Float64},
+                          sum_dep::Array{Float64,1},
+                          sum_nor::Array{Float64,1},
+                          state_dep::Matrix{Float64},
+                          state_nor::Matrix{Float64},
                           allocated::Vector{Bool})
-            FILLLAST!(X::Array{Bool,2},
+        FILLLAST!(X::Array{Bool,2},
                       cap_left::Array{Int64,1},
                       allocated::Vector{Bool})
         end
